fix: progress 값 Long -> Integer로 수정

+++ src/main/java/com/_oormthon/seasonthon/domain/todo/dto/res/TodoResponse.java
@@ -28,7 +28,7 @@ public record TodoResponse(
         String warmMessage,
 
         @Schema(description = "진행률", example = "50")
-        Long progress,
+        Integer progress,
 
         @Schema(description = "Step 리스트")
         List<StepResponse> stepResponses
